# go-gin-app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An on-campus dating application",
        "title": "UFMingle",
        "contact": {},
        "version": "2.0"
    },
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Show forum home page and all articles",
                "responses": {
                    "200": {
                        "description": "Return all the information of article",
                        "schema": {
                            "$ref": "#/definitions/main.article"
                        }
                    }
                }
            }
        },
        "/article/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create an article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The title of the article",
                        "name": "title",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The content of the article",
                        "name": "content",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The author of the article",
                        "name": "author",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If the article has been created successfully, return the number of rows been affected, else 0",
                        "schema": {
                            "type": "int"
                        }
                    },
                    "400": {
                        "description": "There is an error while creating the article",
                        "schema": {
                            "type": "error"
                        }
                    }
                }
            }
        },
        "/article/pastposts/:username": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get article posted by the user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username, i.e. author of the article",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.article"
                            }
                        }
                    },
                    "400": {
                        "description": "failure",
                        "schema": {
                            "type": "error"
                        }
                    }
                }
            }
        },
        "/article/view/:article_id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Open the article page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The index of the article",
                        "name": "article_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the article",
                        "schema": {
                            "$ref": "#/definitions/main.article"
                        }
                    },
                    "404": {
                        "description": "Not found or invalid article_id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/image/avatar/:username": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get the avatar of the user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An avatar is returned",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "error"
                        }
                    }
                }
            }
        },
        "/image/delete/:filename": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Delete an image file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filename of the image",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "map"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "error"
                        }
                    }
                }
            }
        },
        "/image/download/:filename": {
            "get": {
                "produces": [
                    "image/jpeg"
                ],
                "summary": "Retrieve images inserted in the posts or replies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image filename",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "error"
                        }
                    }
                }
            }
        },
        "/image/upload": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Upload images inserted by users in posts or replies",
                "responses": {
                    "200": {
                        "description": "errno: 0, data: A list of download addresses of images",
                        "schema": {
                            "type": "map"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "error"
                        }
                    }
                }
            }
        },
        "/u/login": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Show the login page",
                "responses": {}
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Perform function login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log in successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Failed to log in",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/u/logout": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "Log out successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/u/register": {
            "get": {
                "summary": "Show the registration page",
                "responses": {}
            },
            "post": {
                "summary": "Register a new user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gatorlink",
                        "name": "gatorlink",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "GatorPW",
                        "name": "gatorPW",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register successfully, return the number of rows been affected",
                        "schema": {
                            "type": "int"
                        }
                    },
                    "400": {
                        "description": "ErrorMessage",
                        "schema": {
                            "type": "error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.article": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "dislikes": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}